+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - [Feature] New screen streaming design with animation
 - [Feature] Migrate screen streaming to hub
 - [Feature] Dialog on lock/unlock unsupported by firmware
+- [Feature] Dialog on first pair failed
 - [GitHub] Update github repo picture
 - [CI] Migrate to Detekt
 - [CI] Android lint in CI